requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.2
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
discord.py==1.7.3
idna==3.3
itsdangerous==2.1.2
MarkupSafe==2.1.1
multidict==6.0.2
python-decouple==3.6
requests==2.27.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
